requirements.txt
PyYAML>=6.0
paramiko>=3.0.0